chore: move platform agnostic code in jazz-react-core
