+++ packages/jazz-react-core/.npmignore
@@ -1,0 +1,2 @@
+coverage
+node_modules
