+++ packages/jazz-react-core/README.md
@@ -1,0 +1,5 @@
+# `jazz-react-core`
+
+Platform-agnostic React bindings for [jazz.tools](https://jazz.tools), a framework for distributed state.
+
+Used as building block for [jazz-react-native](https://github.com/garden-co/jazz/tree/main/packages/jazz-react-native) and [jazz-react](https://github.com/garden-co/jazz/tree/main/packages/jazz-react).
