+++ packages/jazz-react-core/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "jazz-react",
+  "name": "jazz-react-core",
   "version": "0.8.45",
   "type": "module",
   "main": "dist/index.js",
@@ -16,11 +16,8 @@
     }
   },
   "dependencies": {
-    "@scure/bip39": "^1.3.0",
     "cojson": "workspace:0.8.45",
-    "jazz-browser": "workspace:0.8.45",
-    "jazz-tools": "workspace:0.8.45",
-    "jazz-react-core": "workspace:*"
+    "jazz-tools": "workspace:0.8.45"
   },
   "devDependencies": {
     "@testing-library/dom": "^10.4.0",
