+++ packages/jazz-react-core/src/index.ts
@@ -1,0 +1,2 @@
+export * from "./hooks.js";
+export * from "./provider.js";
