+++ packages/jazz-react-core/src/tests/testUtils.tsx
@@ -1,0 +1,55 @@
+import {
+  RenderHookOptions,
+  RenderOptions,
+  render,
+  renderHook,
+} from "@testing-library/react";
+import { AnonymousJazzAgent } from "jazz-tools";
+import React from "react";
+import { RegisteredAccount } from "../provider.js";
+import { JazzTestProvider } from "../testing.js";
+
+type JazzExtendedOptions = {
+  account: RegisteredAccount | { guest: AnonymousJazzAgent };
+};
+
+const customRender = (
+  ui: React.ReactNode,
+  options: RenderOptions & JazzExtendedOptions,
+) => {
+  const AllTheProviders = ({
+    children,
+  }: {
+    children: React.ReactNode;
+  }) => {
+    return (
+      <JazzTestProvider account={options.account}>{children}</JazzTestProvider>
+    );
+  };
+
+  return render(ui, { wrapper: AllTheProviders, ...options });
+};
+
+const customRenderHook = <TProps, TResult>(
+  callback: (props: TProps) => TResult,
+  options: RenderHookOptions<TProps> & JazzExtendedOptions,
+) => {
+  const AllTheProviders = ({
+    children,
+  }: {
+    children: React.ReactNode;
+  }) => {
+    return (
+      <JazzTestProvider account={options.account}>{children}</JazzTestProvider>
+    );
+  };
+
+  return renderHook(callback, { wrapper: AllTheProviders, ...options });
+};
+
+// re-export everything
+export * from "@testing-library/react";
+
+// override render method
+export { customRender as render };
+export { customRenderHook as renderHook };
